+++ example/test/boot_cljs_test_example/suite.cljs
@@ -3,4 +3,8 @@
    [doo.runner :refer-macros [doo-tests]]
    [boot-cljs-test-example.lib-test]))
 
+(enable-console-print!)
+
+(println "** custom suite **")
+
 (doo-tests 'boot-cljs-test-example.lib-test)